Replace the kludgy attribute with a system-defined macro.

+++ Source/DOM classes/Core DOM/SVGKCSSPrimitiveValue_ConfigurablePixelsPerInch.h
@@ -2,6 +2,6 @@
 
 @interface SVGKCSSPrimitiveValue ()
 
-@property(nonatomic) float pixelsPerInch;
+@property(nonatomic) CGFloat pixelsPerInch;
 
 @end
